+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 - Add get_qt
+- Add VERSION argument to get_cqtdeployer
 - Remove x86_64-w64-mingw32 toolchain file
 
 ## 0.7.0